requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.41.0
cryptography==44.0.0
dj-rest-auth==7.0.1
Django==5.1.4
django-allauth==65.3.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
razorpay==1.4.2
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
